+++ patches/src/main/kotlin/hoodles/morphe/patches/ibispaint/prime/EnablePrimePatch.kt
@@ -14,7 +14,7 @@ val enablePrimePatch = rawResourcePatch(
         name = "IbisPaint X",
         packageName = "jp.ne.ibis.ibispaintx.app",
         appIconColor = 0x0B649B,
-        targets = listOf(AppTarget("14.0.1"))
+        targets = listOf(AppTarget("14.0.4"))
     ))
 
     dependsOn(
@@ -22,10 +22,10 @@ val enablePrimePatch = rawResourcePatch(
             val libPath = "lib/arm64-v8a/libibispaint.so"
             
             // disable anti-tamper
-            "28 AE 00 90 08 55 47 F9 08 FD DF 88 1F 05 00 71" asPatternTo "28 AE 00 90 08 55 47 F9 E0 03 1F 2A C0 03 5F D6" inFile libPath
+            "08 AF 00 B0 08 D9 47 F9 08 FD DF 88 1F 05 00 71" asPatternTo "E0 03 1F 2A C0 03 5F D6 08 FD DF 88 1F 05 00 71" inFile libPath
 
             // enable prime membership
-            "4A 4E 46 94 00 00 62 9E 00 21 60 1E E0 B7 9F 1A 02 00 00 14 E0 03 1F 2A" asPatternTo "4A 4E 46 94 00 00 62 9E 00 21 60 1E E0 B7 9F 1A 1F 20 03 D5 20 00 80 52" inFile libPath
+            "45 39 47 94 00 00 62 9E 00 21 60 1E E0 B7 9F 1A 02 00 00 14 E0 03 1F 2A" asPatternTo "45 39 47 94 00 00 62 9E 00 21 60 1E E0 B7 9F 1A 1F 20 03 D5 20 00 80 52" inFile libPath
         })
     )
 }