+++ patches/src/main/kotlin/hoodles/morphe/patches/xodo/misc/signature/DisableSignatureCheckPatch.kt
@@ -4,12 +4,7 @@ import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.util.returnEarly
 import hoodles.morphe.patches.xodo.shared.Constants
 
-val disableSignatureCheckPatch = bytecodePatch(
-    name = "Disable signature check",
-    description = "Removes the anti-tamper protection, which verifies apk signature, causing the app to force close."
-) {
-    compatibleWith(Constants.COMPATIBILITY)
-
+val disableSignatureCheckPatch = bytecodePatch{
     execute {
         SignatureCheckFingerprint.method.returnEarly()
     }