+++ patches/src/main/kotlin/hoodles/morphe/patches/xodo/misc/signature/Fingerprints.kt
@@ -3,5 +3,6 @@ package hoodles.morphe.patches.xodo.misc.signature
 import app.morphe.patcher.Fingerprint
 
 object SignatureCheckFingerprint : Fingerprint(
-    strings = listOf("completeReaderMainActivity", "completeReaderMainActivity.application")
+    parameters = listOf("Lviewer/CompleteReaderMainActivity;"),
+    strings = listOf("UTF_8")
 )